Remove ISymbolService.GetMetaData() and replace with IModuleService.CreateModule call
Move/rename ManagedImageMappingModuleService into ImageMappingMemoryService class
