+++ src/Microsoft.Diagnostics.DebugServices/IModuleService.cs
@@ -1,6 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System;
 using System.Collections.Generic;
 
 namespace Microsoft.Diagnostics.DebugServices
@@ -47,13 +48,14 @@ namespace Microsoft.Diagnostics.DebugServices
         IEnumerable<IModule> GetModuleFromModuleName(string moduleName);
 
         /// <summary>
-        /// Create a module instance from a stream (memory or file).
+        /// Create a module instance
         /// </summary>
-        /// <param name="moduleIndex">artifical index</param>
+        /// <param name="moduleIndex">artificial index or -1 for none</param>
         /// <param name="imageBase">module base address</param>
         /// <param name="imageSize">module size</param>
         /// <param name="imageName">module name</param>
         /// <returns>IModule</returns>
+        /// <exception cref="ArgumentNullException">thrown if imageBase or imageSize is 0</exception>
         IModule CreateModule(int moduleIndex, ulong imageBase, ulong imageSize, string imageName);
     }
 }
